--- src/Core/tests/DeviceTests.iOS/GlobalSuppressions.cs
@@ -1,6 +1,0 @@
-// This file is used by Code Analysis to maintain SuppressMessage
-// attributes that are applied to this project.
-// Project-level suppressions either have no target or are given
-// a specific target and scoped to a namespace, type, member, etc.
-
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.NamingRules", "SA1300:Element should begin with upper-case letter", Justification = "iOS is what we want.", Scope = "namespace", Target = "~N:Microsoft.Maui.DeviceTests.iOS")]
