+++ src/Core/tests/DeviceTests.iOS/Main.cs
@@ -1,16 +1,16 @@
-using System;
+using Microsoft.Maui.TestUtils;
 using UIKit;
 
 namespace Microsoft.Maui.DeviceTests
 {
-    public class Application
-    {
-        static void Main(string[] args)
-        {
-            if (args?.Length > 0 || Environment.GetEnvironmentVariable("NUNIT_AUTOEXIT")?.Length > 0) // usually means this is from xharness
-                UIApplication.Main(args, null, nameof(TestApplicationDelegate));
-            else
-                UIApplication.Main(args, null, nameof(AppDelegate));
-        }
-    }
+	public class Application
+	{
+		static void Main(string[] args)
+		{
+			if (BaseTestApplicationDelegate.IsXHarnessRun(args))
+				UIApplication.Main(args, null, nameof(TestApplicationDelegate));
+			else
+				UIApplication.Main(args, null, nameof(AppDelegate));
+		}
+	}
 }
