+++ src/Essentials/samples/Samples.Android/Resources/values/Strings.xml
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <resources>
-  <string name="app_name">Xamarin.Essentials</string>
+  <string name="app_name">Essentials</string>
 </resources>
