+++ src/Essentials/samples/Samples.UWP/Package.appxmanifest
@@ -3,7 +3,7 @@
   <Identity Name="ec0cc741-fd3e-485c-81be-68815c480690" Publisher="CN=Microsoft" Version="1.0.0.0" />
   <mp:PhoneIdentity PhoneProductId="ec0cc741-fd3e-485c-81be-68815c480690" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
   <Properties>
-    <DisplayName>Xamarin.Essentials</DisplayName>
+    <DisplayName>Essentials</DisplayName>
     <PublisherDisplayName>Microsoft</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -15,7 +15,7 @@
   </Resources>
   <Applications>
     <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="Samples.UWP.App">
-      <uap:VisualElements DisplayName="Xamarin.Essentials" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="Samples" BackgroundColor="transparent">
+      <uap:VisualElements DisplayName="Essentials" Square150x150Logo="Assets\Square150x150Logo.png" Square44x44Logo="Assets\Square44x44Logo.png" Description="Samples" BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png">
         </uap:DefaultTile>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
