+++ src/Essentials/samples/Samples.iOS/Info.plist
@@ -24,7 +24,7 @@
 	<key>MinimumOSVersion</key>
 	<string>10.0</string>
 	<key>CFBundleDisplayName</key>
-	<string>Xamarin.Essentials</string>
+	<string>Essentials</string>
 	<key>CFBundleIdentifier</key>
 	<string>com.microsoft.maui.essentials</string>
 	<key>CFBundleVersion</key>
@@ -32,7 +32,7 @@
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
 	<key>CFBundleName</key>
-	<string>Xamarin.Essentials</string>
+	<string>Essentials</string>
 	<key>CFBundleShortVersionString</key>
 	<string>1.0</string>
 	<key>NSLocationWhenInUseUsageDescription</key>
