+++ src/Essentials/src/nugetreadme.txt
@@ -1,11 +1,11 @@
-## Xamarin.Essentials
+## Microsoft.Maui.Essentials
 
-Thank you for installing Xamarin.Essentials, be sure to read through our full documentation at:
+Thank you for installing Microsoft.Maui.Essentials, be sure to read through our full documentation at:
 https://aka.ms/xamarinessentials
 
 ## Setup
 
-Ensure that you install Xamarin.Essentials into all of your projects.
+Ensure that you install Microsoft.Maui.Essentials into all of your projects.
 
 For Android projects there is a small amount of setup needed to handle permissions. Please follow our short guide at:
 https://aka.ms/essentials-getstarted
