+++ src/Essentials/test/DeviceTests.Android/Essentials.DeviceTests.Android.csproj
@@ -64,6 +64,10 @@
     <PackageReference Include="Xamarin.AndroidX.Palette" Version="1.0.0.6" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\..\TestUtils\TestUtils.DeviceTests\src\TestUtils.DeviceTests.csproj">
+      <Project>{551b2209-4298-4d60-b55c-79077b8bc244}</Project>
+      <Name>TestUtils.DeviceTests</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\src\Essentials.csproj">
       <Project>{8CB95D25-8442-42BC-82BE-319D21138549}</Project>
       <Name>Essentials</Name>
