+++ src/Essentials/test/DeviceTests.Android/MainActivity.cs
@@ -6,31 +6,31 @@ using Xunit.Runners.UI;
 
 namespace DeviceTests.Droid
 {
-    [Activity(
-        Name = "com.microsoft.maui.essentials.devicetests.MainActivity",
-        Label = "@string/app_name",
-        Theme = "@style/MainTheme",
-        MainLauncher = true,
-        ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation)]
-    public class MainActivity : RunnerActivity
-    {
-        protected override void OnCreate(Bundle bundle)
-        {
-            Microsoft.Maui.Essentials.Platform.Init(this, bundle);
+	[Activity(
+		Name = "com.microsoft.maui.essentials.devicetests.MainActivity",
+		Label = "@string/app_name",
+		Theme = "@style/MainTheme",
+		MainLauncher = true,
+		ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation)]
+	public class MainActivity : RunnerActivity
+	{
+		protected override void OnCreate(Bundle bundle)
+		{
+			Microsoft.Maui.Essentials.Platform.Init(this, bundle);
 
-            // tests can be inside the main assembly
-            AddTestAssembly(Assembly.GetExecutingAssembly());
-            AddTestAssembly(typeof(Battery_Tests).Assembly);
-            AddExecutionAssembly(typeof(Battery_Tests).Assembly);
+			// tests can be inside the main assembly
+			AddTestAssembly(Assembly.GetExecutingAssembly());
+			AddTestAssembly(typeof(Battery_Tests).Assembly);
+			AddExecutionAssembly(typeof(Battery_Tests).Assembly);
 
-            base.OnCreate(bundle);
-        }
+			base.OnCreate(bundle);
+		}
 
-        public override void OnRequestPermissionsResult(int requestCode, string[] permissions, Permission[] grantResults)
-        {
-            Microsoft.Maui.Essentials.Platform.OnRequestPermissionsResult(requestCode, permissions, grantResults);
+		public override void OnRequestPermissionsResult(int requestCode, string[] permissions, Permission[] grantResults)
+		{
+			Microsoft.Maui.Essentials.Platform.OnRequestPermissionsResult(requestCode, permissions, grantResults);
 
-            base.OnRequestPermissionsResult(requestCode, permissions, grantResults);
-        }
-    }
+			base.OnRequestPermissionsResult(requestCode, permissions, grantResults);
+		}
+	}
 }
